feat: Dibs 기능 및 관련 DTO, 서비스, 컨트롤러 구현
- DibsController 및 DibsService를 추가하여 찜하기 기능을 구현
- DibsDto를 생성하여 찜하기 응답 형식을 정의
- StudyGroupController 및 관련 서비스에서 ID 타입을 Integer에서 Long으로 변경하여 일관성 유지
- DibsRepository 및 관련 테스트 케이스 추가
- 기존 코드에서 Integer ID를 Long으로 변경하여 데이터베이스와의 호환성 향상
- DibsService의 toggleDibs 메서드에서 찜하기 추가 및 삭제 로직 구현
- 관련 테스트 케이스를 통해 기능 검증 및 안정성 확보

+++ src/main/java/com/depth/learningcrew/domain/studygroup/controller/StudyGroupApplicationController.java
@@ -34,7 +34,7 @@ public class StudyGroupApplicationController {
     @GetMapping("/{groupId}/applications")
     @Operation(summary = "스터디 그룹 가입 신청 목록 조회", description = "스터디 그룹의 owner가 가입 신청 목록을 페이지네이션으로 조회합니다.")
     public PagedModel<ApplicationDto.ApplicationResponse> getApplicationsByGroupId(
-            @Parameter(description = "스터디 그룹 ID") @PathVariable Integer groupId,
+            @Parameter(description = "스터디 그룹 ID") @PathVariable Long groupId,
             @ModelAttribute @ParameterObject ApplicationDto.SearchConditions searchConditions,
             @PageableDefault @ParameterObject Pageable pageable,
             @Parameter(hidden = true) @AuthenticationPrincipal UserDetails userDetails) {
@@ -47,7 +47,7 @@ public class StudyGroupApplicationController {
     @Operation(summary = "스터디 그룹 가입 신청", description = "로그인한 사용자가 특정 스터디 그룹에 가입 신청을 합니다.")
     @ResponseStatus(HttpStatus.CREATED)
     public ApplicationDto.ApplicationResponse joinStudyGroup(
-            @Parameter(description = "스터디 그룹 ID") @PathVariable Integer groupId,
+            @Parameter(description = "스터디 그룹 ID") @PathVariable Long groupId,
             @Parameter(hidden = true) @AuthenticationPrincipal UserDetails userDetails) {
 
         return studyGroupApplicationService.joinStudyGroup(groupId, userDetails);
@@ -56,8 +56,8 @@ public class StudyGroupApplicationController {
     @PostMapping("/{groupId}/applications/{userId}/approve")
     @Operation(summary = "스터디 그룹 가입 신청 수락", description = "스터디 그룹의 owner가 가입 신청을 수락합니다.")
     public ApplicationDto.ApplicationResponse approveApplication(
-            @Parameter(description = "스터디 그룹 ID") @PathVariable Integer groupId,
-            @Parameter(description = "신청자 ID") @PathVariable Integer userId,
+            @Parameter(description = "스터디 그룹 ID") @PathVariable Long groupId,
+            @Parameter(description = "신청자 ID") @PathVariable Long userId,
             @Parameter(hidden = true) @AuthenticationPrincipal UserDetails userDetails) {
 
         return studyGroupApplicationService.approveApplication(groupId, userId, userDetails);
@@ -66,8 +66,8 @@ public class StudyGroupApplicationController {
     @PostMapping("/{groupId}/applications/{userId}/reject")
     @Operation(summary = "스터디 그룹 가입 신청 거절", description = "스터디 그룹의 owner가 가입 신청을 거절합니다.")
     public ApplicationDto.ApplicationResponse rejectApplication(
-            @Parameter(description = "스터디 그룹 ID") @PathVariable Integer groupId,
-            @Parameter(description = "신청자 ID") @PathVariable Integer userId,
+            @Parameter(description = "스터디 그룹 ID") @PathVariable Long groupId,
+            @Parameter(description = "신청자 ID") @PathVariable Long userId,
             @Parameter(hidden = true) @AuthenticationPrincipal UserDetails userDetails) {
 
         return studyGroupApplicationService.rejectApplication(groupId, userId, userDetails);
